Migrate mod database tRPC router to use content database
